3.1.2 stability - Use tab character for indentation - Accessibility enhancements on path navigator - Fix typo/spelling
