+++ prettier.config.js
@@ -17,7 +17,7 @@
 module.exports = {
   printWidth: 100,
   tabWidth: 2,
-  useTabs: false,
+  useTabs: true,
   semi: true,
   singleQuote: true,
   quoteProps: 'as-needed',
